requirements.txt
numpy
scikit-image
torch
torchvision
pillow
opencv-python-headless
paddlepaddle
paddlehub

pyyaml
tqdm
numpy
easydict==1.9.0
scikit-image==0.17.2
scikit-learn==0.24.2
tensorflow
joblib
matplotlib
pandas
albumentations==0.5.2
hydra-core==1.1.0
pytorch-lightning==1.2.9
tabulate
easydict
kornia==0.5.0
webdataset
packaging
scikit-learn==0.24.2
